ROX-34952: address review feedback - use set.Set, separate file, single loop
