+++ pkg/booleanpolicy/filter/compile.go
@@ -20,41 +20,3 @@ func CompileEvaluationFilter(proto *storage.EvaluationFilter) []EvaluationFilter
 	}
 	return filters
 }
-
-func newContainerTypeFilter(skipTypes []storage.ContainerType) *EvaluationFilter {
-	if len(skipTypes) == 0 {
-		return nil
-	}
-	skip := make(map[storage.ContainerType]struct{}, len(skipTypes))
-	for _, t := range skipTypes {
-		skip[t] = struct{}{}
-	}
-	return &EvaluationFilter{
-		isNonDefault: func() bool { return true },
-		apply: func(dep *storage.Deployment, imgs []*storage.Image) (*storage.Deployment, []*storage.Image) {
-			hasSkipped := false
-			for _, c := range dep.GetContainers() {
-				if _, ok := skip[c.GetType()]; ok {
-					hasSkipped = true
-					break
-				}
-			}
-			if !hasSkipped {
-				return dep, imgs
-			}
-			filtered := dep.CloneVT()
-			filtered.Containers = nil
-			var filteredImgs []*storage.Image
-			for i, c := range dep.GetContainers() {
-				if _, ok := skip[c.GetType()]; ok {
-					continue
-				}
-				filtered.Containers = append(filtered.Containers, c)
-				if i < len(imgs) {
-					filteredImgs = append(filteredImgs, imgs[i])
-				}
-			}
-			return filtered, filteredImgs
-		},
-	}
-}
